Create validatepop package
Creates validatepop package in model_code with a couple of populatoin
validation functions, tests etc. From the package DESCRIPTION:

Functions that perform common checks on demographic dataset,
throwing errors when the inputs are not what is expected.
Currently contains two functions, validate_population(),
which checks that a dataset conforms to a set of expectations about a population dataset,
and validate_join_population() which checks that one population's aggregation levels
are contained within another's and that a join will behave as expected.