Add generalpop package with popn_apply_rate function
Plus something of a reorganisation
- the newgeneralpop package will be for functions that are reused
across the codebase
- functions that were in general/ have been moved to generalpop/R/,
as the name 'general' is too, uh, general for a package
- the function popn_apply_rate applies a rate to a population. It was
created from the deaths_from_popn_mort function in the deaths package
with almost no changes except to variable names, and removing a
couple of restrictions on the input
- the births_from_popn_fert and deaths_from_popn_mort functions in the
births and deaths packages have been updated to work with
popn_apply_rate, including removing a number of standardisation things
that are now done inside popn_apply_rate. The checks and tests are
unchanged, and maybe a little excessive now.