popn_age_on - improve, test and integrate to trend model
Add flexibility to popn_age_on - it can now work with custom timesteps,
and with age bands (i.e. as factors), and conducts various data
validations on the input and output. Testing and documentation also
added, and the trend model updated to work with it. Only small impact on
speed.