Improve validation speed + bugfix
With the new profvis package, there were a couple of things that were
obviously eating time in dataset validation. A couple of the big ones
were fixed (but there's still space for improvement!)

- Formerly the validation function converted all aggregation levels to
factors to enable easier comparisons and calculations under the mistaken
assumption that this would speed things up. It really didn't, and now
factor conversion is only performed when comparing to a template data
frame with a different factoring structure.

- The comparison of two populations takes some time. There is a minor
speed boost if we can assume that the both are complete in their
aggregation levels (already set as a parameter), in which case we only
need to check that the levels are the same in each population.

Also, a bugfix in the (oft-repeated) match_factors() function, which
should be split out into a single location - it would fail to recognise
an integer as numeric. This has been fixed.