Cf better warnings (#72)
* Reduce warnings

Correct one test for rates between 0 and 1 in 02_core.R that was testing
a non-existent column

Remove a warning when migrate domestic doesn't have rates for an
aggregation level (and leave a TODO to replace it with something else)

* Improve model warnings

- In each loop of the model core, identify negative populations created
and write out a summary data frame for them in a warning message before
they're set to zero

- When creating the domestic migration input identify total outmigration
rates exceeding a cap rate (0.8 in the model). Print a summary table of
locations exceeding the limit in a warning message before rescaling the
rates to sum to 0.8. (Previously the rates were capped at 0.8 for each
origin and each destination, which wasn't how we wanted to do this.)

* Bugfix international migration validation fn