Bugfix recode_gss_codes
Using data.table in recode_gss_to_2011 meant that it would modify the
input data by reference if there was no extra aggregation needed at the
end. This was potentially dangerous, so the code takes a copy of the
input before it starts. Benchmarking shows that there's no extra cost.