Bugfix SA dwelling trajectory, update backseries
The small area model was giving weird dwelling trajectories because the
cumulative sum of annual changes in units was being performed on a data
frame in the wrong order.

Also now that the housing led model outputs backseries, there's no need
to add backseries to the output - I kept the code and filtered for years
< 2011, which sometimes exist.

Also see next commit for bugfix to fertility rates - the
small_area_fertility_rates data frame contains fewer columns, which was
the source of some confusion.