Define global variables in global.R
global.R is a way of avoiding a million NOTEs in the R CMD CHECK output
(or devtools::check() output).

The check complains when it sees variables that it hasn't already seen
defined, and unfortunately the non-standard evaluation of the tidyverse
falls foul of this - e.g.
  select(mtcars, cyl, mpg)
will complain that 'cyl' and 'mpg' aren't defined. global.R is a huge
list of exceptions that the check scripts can see.

Created following instructions here
https://www.r-bloggers.com/no-visible-binding-for-global-variable/