Clarify trend model output dir
The trend model worked with two variables - 'outputs_dir', defined in
the config file, and 'output_dir', defined in the control. The only
purpose of outputs_dir was to create the path to output_dir, and the
other models just use output_dir and define it in the config.

This makes a few changes so to replace outputs_dir with output_dir, and
drops a couple of checks that are no longer necessary.

Note: this DOESN'T add a datestamp to the output dir, which other
commits in this branch do for other models.