Fix dtplyr warning in popn_age_on
This was a tricky one. Data table functionality in the population age on
function was converting the input population in the parent environment
to a data table, but conversions back to data frame at the end of the
function weren't operating on the copy in the parent environment,
because 'mutate'-style functions in the meantime meant that a copy of
the population had been taken and it was no longer a pointer to the
parent environment.

Cleaned it up by taking a copy and assigning it to the 'aged' variable
slightly earlier in the calculations and limiting all data.table
operations to this copy within the function.