Update gss lookups to include regions
This means updating the lookup file to a new one on the Q drive which is
picked up by the initialize script

It's fixing a problem where some outputs didn't have regional names in
them.

Also used get_gss_names() to get lookup data instead of reading the file
directly for consistency.