Bugfix domestic and regional aggregations
The domestic migration backseries data was being wrongly aggregated.
Removed the code that does that and created a Jira ticket to fix it.

Edits to the model core domestic data aggregations so that
all-England is included.

Non-fix to the household models where inputs are filtered to remove
regional aggregations because they were causing trouble.