Hunting down dtplyr warnings + misc
Running the initialize scripts gave loads of warnings from dtplyr where
tidyverse methods were used on data frames. This converts things to
tibbles or data frames to avoid the messages. Tibbles are used because
data frames can't handle column names that start with numbers (typically
years in the input data) and add a preceding 'X' when faced with them.

Also, add a check to the missing data interpolation in mye_2018 so that
it fails if it tries to interpolate across groupings (which it doesn't
currently need to). Easier than fixing it.