Change recode_gss_code aggregation
Let's talk about this - recode_gss_codes iterates through a large number
of renames and merges, each time updating gss_codes in a data frame. At
the end it calculates sums/means over any duplicated codes.

This change edits things so that it sums over any duplicated codes after
each update to the codes. As far as I can tell this ordering only
matters when means are being taken, and I don't know if it's the right
approach.

If we want to keep the aggregation at the end, we could speed up the
function considerable by collapsing all the recodes into a single
vector that gets the multiple renaming/recodings correct. Let's discuss.