Don't output domestic rates in trend, fix checks
With the new domestic rates methods it's possible for the rates to
change many times during a model run. Domestic rates are too huge to
write out for every year so they're no longer part of the output (this
is the origin-destination rates to be clear).

Also fixed some of the model validation checks.