Changes to lock file so it works on azure
Weird things happen when trying to install our population packages when
it tries to test it on a 32 bit architecture.  It complains about not
finding a DLL file for certain packages.  Doing renv::install("package",
type = "binary") seems to solve the problem.  I don't know why the
previous version worked on the surface pro but not the azure box

Also added our 4 population models to the ignore setting for
renv::snapshot